Move services to their own folder. Add typings for Google libraries to typings.json.
