+++ src/client/app/service/google-realtime/index.ts
@@ -1,0 +1,4 @@
+/**
+ * This barrel file provides the export for the shared NameListService.
+ */
+export * from './google-realtime.service';
