+++ src/client/app/service/name-list/name-list.service.ts
