+++ src/client/app/shared/index.ts
@@ -1,7 +1,6 @@
 /**
  * This barrel file provides the exports for the shared resources (services, components).
  */
-export * from './name-list/index';
 export * from './graph-preview-list/index';
 export * from './side-panel/index';
 export * from './config/env.config';
