+++ tools/tasks/seed/build.js.prod.ts
@@ -14,9 +14,9 @@ const INLINE_OPTIONS = {
 };
 
 /**
- * Executes the build process, transpiling the TypeScript files for the production environment.
+ * Executes the build process, transpiling the TypeScript files for the
+ * production environment.
  */
-
 export = () => {
   let tsProject = makeTsProject();
   let src = [
