+++ typings.json
@@ -12,7 +12,10 @@
     "chokidar": "github:DefinitelyTyped/DefinitelyTyped/chokidar/chokidar.d.ts#ae81e340a9e897167cfef3122a5a3aed04293814",
     "core-js": "registry:dt/core-js#0.0.0+20160725163759",
     "express": "github:DefinitelyTyped/DefinitelyTyped/express/express.d.ts#5c3e47967affa3c4128a3875d1664ba206ae1b0f",
+    "gapi": "registry:dt/gapi#0.0.0+20160412153822",
+    "gapi.auth2": "registry:dt/gapi.auth2#0.0.0+20160602144801",
     "glob": "github:DefinitelyTyped/DefinitelyTyped/glob/glob.d.ts#5c3e47967affa3c4128a3875d1664ba206ae1b0f",
+    "google-drive-realtime-api": "registry:dt/google-drive-realtime-api#0.0.0+20161012131800",
     "gulp": "github:DefinitelyTyped/DefinitelyTyped/gulp/gulp.d.ts#5c3e47967affa3c4128a3875d1664ba206ae1b0f",
     "gulp-load-plugins": "github:DefinitelyTyped/DefinitelyTyped/gulp-load-plugins/gulp-load-plugins.d.ts#e081148d88b857d66509e3b46edbd08b3f75f96a",
     "gulp-protractor": "registry:dt/gulp-protractor#1.0.0+20160316155526",
